requirements.txt
Flask==2.1.0
google-cloud-storage>=1.17.0